Add preprocessing step to insert BedMachine thk into gridded dataset
The reference gridded dataset has a funky thickness field that includes
non-ice-sheet thin ice along the Antarctic Peninsula and is inconsistent
with the thickness field we ultimately interpolate from using the
BedMachine dataset.  This leads to unusual gutter sizes in parts of the
mesh and may be contributing to some strange culling artifacts.

This commit adds a preprocessing function that replaces the reference
gridded dataset's thickness field with one bilinearly interpolated from
BedMachine.  This is only used for the flood fill and mesh culling steps.

Co-authored-by: Trevor Hillebrand <trhille@lanl.gov>