Scale fluxes so that total mass is conserved
Since the original mass is on the plane, fluxes must be scaled
by the ratio of the area on the plan to the area on the sphere
to ensure that the remapped flux has the expected sum.

For good measure, this merge also rescales the flux after remapping
using the area on the sphere over MPAS's `areaCell` (which are
very close).