Remove `vertex` masks from defaults of region mask function.
The vertex (and Edge) masks slow down the `antarctic/mesh_gen` testcase
considerably (~50% for default config options). As _very_ few fields
are defined on vertices, we've opted to remove vertex masks from the
defaults and only include cell and edge maks, as those are where
variables are more commonly defined.