Only make tidal mixing files for meshes with cavities
Since the tidal model used to produce the dataset only covers the
Southern Ocean, there is not much value in having the reampped
dataset for meshes without cavities.  If we want to make a more
general mixing parameterization that uses tidal information, we
would need to use data from a global tidal model for the input
dataset in the future.