Add config options for min levels
Previously, we specified the minimum thickness of the global ocean,
which was set to 10 m.

This is not a good approach with different vertical coordinates.  If
we want to use the same settings across various vertical coordinates,
it is more useful to require a minimum number of levels.

This merge changes the global ocean `initial_state` step to use
config options to set the number of vertical levels both in the
open ocean and in ice-shelf cavities.  Relatedly, we also add a
config option to set the minimum layer thickness in cavities.

This change will affect the vertical 3D vertical cooridinate in
all meshes (hopefully for the better).

For meshes that use variants of the 80-layer vertical coordinate,
we set the minimum number of layers that sets the column thickness
to either 10 or 20 m.