Update geojson and mesh parameters
Use a custom geojson that extends the gutter where we expect many
runs to predict margin advance. Also update mesh parameters for a
high resolution mesh that uses only bed topograhpy to set cell spacing,
and fix bug that was setting cell spacing to the maximum value
outside of the ice mask.