Add a step for smoothing topograpy
We remap the topography once without smoothing and then once
with smoothing.  The unsmoothed topography is used to determine
the land mask, thus allowing many meshes with different levels
of smoothing to have the same mapping files in E3SM.

For the time being, only two simple types of smoothing are supported:
a smoothing over a fixed distance or by expanding cells by a fixed
fraction.  (The two types of smoothing can also be used at the
same time.)

In the future, maps with non-uniform smoothing could be added for
finer control.

A check use used to bypass the second smoothing pass if, in fact,
no smoothing is requested (the default behavior).