Create option for high res in ocean for gis mesh
This commit creates the option to maintain high resolution in all ocean
cells (same resolution as what is prescribed at outlet glacier termini).
This is necessary to resolve the fjord geometry that thermal forcing
parameterizations rely on. To reduce time when generating these higher
resolution meshes, this commit also introduces an option to do an
initial cull of the bedMachine dataset using the extent of the
contintental shelf before generating the GIS mesh.